feat: add `baseUrl` option (#23)
To allow specifying the path of the JS files relative to the HTML loading it. This is needed because the preload happens through link elements in the HTML and all paths become relative to the HTML. The `baseUrl` is used in this case to adjust the link href-s.

Fixes #17, #22.

Also fixes #22 by correctly setting `crossOrigin` property in the preload code instead of `crossorigin` (which is the attribute name).